requirements_m2_gpr.txt
numpy==1.26.4
torch>=2.1.0
faiss-cpu>=1.7.4
spacy==3.7.2
transformers>=4.40.1
sentence-transformers>=2.5.1
